leaderboard: Add year-cumulative-sum
Generates a table showing the total points of the top N for a year.

Can be used to show distribution of points

I must have written this code for the third time, and the previous two
times I must have found it so trivial that I didn't bother saving it.
This time I finally got fed up with rewriting the same tired code every
time.