test: Handle .argv and .in together
Over the years there may have been a number of inputs that only make
sense for one part or the other.

Examples:

2023 day 1 (part 2 input doesn't conform to format of part 1)
2020 day 14 (part 1 has too many Xs for part 2 to be reasonable)
2017 day 10 (part 2 inputs are arbitrary strings instead of integers)

In such cases, I'd like to pass a -1 or -2 flag in addition to the input
file, so the test runner needs to support this.

Another use case is when the instructions also say to run the example
input with a different parameter, and this parameter is not specified in
the input. This includes:

2015 day 14
2016 day 08
2016 day 11
2016 day 13
2016 day 16
2016 day 18
2016 day 21
2017 day 16
2017 day 21
2018 day 06
2018 day 07
2020 day 09